+++ tests/test_kubealias.py
@@ -3,17 +3,17 @@ import sys
 import unittest
 
 
-# Ensure the project root is on sys.path so kubealias can be imported when tests
+# Ensure the project root is on sys.path so ``main`` can be imported when tests
 # are run from different working directories.
 ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 if ROOT_DIR not in sys.path:
     sys.path.insert(0, ROOT_DIR)
 
-import kubealias
+import main
 
 
 class TestKubealias(unittest.TestCase):
-    """Unit tests for the :func:`kubealias.create_command` function."""
+    """Unit tests for the :func:`main.create_command` function."""
 
     basic_pairs = {
         "gepo": "kubectl get pods",
@@ -35,18 +35,18 @@ class TestKubealias(unittest.TestCase):
         for alias, expected in self.basic_pairs.items():
             with self.subTest(alias=alias):
                 self.assertEqual(
-                    kubealias.create_command(["prog", alias]),
+                    main.create_command(["prog", alias]),
                     expected,
                 )
 
     def test_help_shown_when_no_alias_provided(self) -> None:
         """When no alias is given, the help message should be returned."""
-        result = kubealias.create_command(["prog"])
+        result = main.create_command(["prog"])
         self.assertIn("Usage:", result)
 
     def test_unknown_alias_returns_message(self) -> None:
         """An unknown alias should return an informative error message."""
-        result = kubealias.create_command(["prog", "gezz"])
+        result = main.create_command(["prog", "gezz"])
         self.assertEqual(
             result,
             "echo \"Alias not found: 'zz'. Call the script without arguments for help.\"",
@@ -54,7 +54,7 @@ class TestKubealias(unittest.TestCase):
 
     def test_ff_alias_builds_fzf_command(self) -> None:
         """The special ``ff`` alias should inject the fzf command snippet."""
-        result = kubealias.create_command(["prog", "geffpo"])
+        result = main.create_command(["prog", "geffpo"])
         snippet = "$(kubectl get pods | fzftab | awk '{print $1}')"
         self.assertIn(snippet, result)
         self.assertTrue(result.startswith("kubectl get "))
